+++ presets/plugins.ts
@@ -1,6 +1,7 @@
 export * from '@groupby/elements-cache-plugin';
 export * from '@groupby/elements-cache-driver-plugin';
 export * from '@groupby/elements-dom-events-plugin';
+export * from '@groupby/elements-gb-tracker-plugin';
 export * from '@groupby/elements-sayt-driver-plugin';
 export * from '@groupby/elements-sayt-plugin';
 export * from '@groupby/elements-search-driver-plugin';