+++ packages/@groupby/elements-sayt-driver-plugin/CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] [patch]
+## [Unreleased] [minor]
+### Added
+- ELE-249: Added basic beaconing functionality.
+  - Added a call to trigger a search beacon after a successful products search.
+  - Added `gb-tracker-client` as a dependency.
+
 ### Fixed
 - ELE-247: Exported the previously hidden `SaytDriverOptions` interface.